feat: terraform IaC template
Created a terraform IaC template, this one creates a VPC, 1 Private Subnet, 1 Public Subnet, 1 IGW for the Public subnet, Security Groups and a EC2 for the Public Subnet.

Also, there is a file called variables.tf that defines the variable for the key pair name(Remember to add a terrafor.tfvars were you create the value for this) and also a provided.tf with the information of the provider.